compose/requirements.txt
fastapi[standard]==0.112.1
uvicorn==0.30.6
pillow==10.4.0
python-multipart==0.0.9
psycopg2-binary==2.9.9
minio==7.2.8
requests==2.32.3
